// src/enums.ts
export enum Currency {
  USD = 1,
  GBP,
  EUR,
  CHF,
  RUB,
  PLN,
  BRL,
  JPY,
  NOK,
  IDR,
  MYR,
  PHP,
  SGD,
  THB,
  VND,
  KRW,
  TRY,
  UAH,
  MXN,
  CAD,
  AUD,
  NZD,
  CNY,
  INR,
  CLP,
  PEN,
  COP,
  ZAR,
  HKD,
  TWD,
  SAR,
  AED,
  ARS = 34,
  ILS,
  KZT = 37,
  KWD,
  QAR,
  CRC,
  UYU
}

export enum Languages {
  brazilian = 'brazilian',
  bulgarian = 'bulgarian',
  czech = 'czech',
  danish = 'danish',
  dutch = 'dutch',
  english = 'english',
  finnish = 'finnish',
  french = 'french',
  german = 'german',
  greek = 'greek',
  hungarian = 'hungarian',
  italian = 'italian',
  japanese = 'japanese',
  koreana = 'koreana',
  latam = 'latam',
  norwegian = 'norwegian',
  polish = 'polish',
  portuguese = 'portuguese',
  romanian = 'romanian',
  russian = 'russian',
  schinese = 'schinese',
  spanish = 'spanish',
  swedish = 'swedish',
  tchinese = 'tchinese',
  thai = 'thai',
  turkish = 'turkish',
  ukrainian = 'ukrainian',
  vietnamese = 'vietnamese'
}

export enum CountryCode {
  BG = 'BG',
  BR = 'BR',
  CN = 'CN',
  CZ = 'CZ',
  DE = 'DE',
  DK = 'DK',
  EN = 'EN',
  ES = 'ES',
  FI = 'FI',
  FR = 'FR',
  GR = 'GR',
  HU = 'HU',
  IT = 'IT',
  JP = 'JP',
  KR = 'KR',
  NL = 'NL',
  NO = 'NO',
  PL = 'PL',
  PT = 'PT',
  RO = 'RO',
  RU = 'RU',
  SE = 'SE',
  TH = 'TH',
  TR = 'TR',
  TW = 'TW',
  UA = 'UA',
  VN = 'VN'
}